feat: video{Resource,Instance}両方から動画プレイヤーが表示できるように
